move profile page into its own sub components
