+++ app/api/signers/stats/route.ts
@@ -1,14 +1,17 @@
 import { NextRequest } from 'next/server'
-import { snapchain, SnapchainCastsResponse, SnapchainReactionsResponse, SnapchainLinksResponse, SnapchainVerificationsResponse, SnapchainMessage } from '../../../../lib/snapchain'
-import { CACHE_TTLS } from '../../../../lib/utils'
+import { snapchain, SnapchainCastsResponse, SnapchainReactionsResponse, SnapchainLinksResponse, SnapchainVerificationsResponse, SnapchainMessage } from '../../../lib/snapchain'
+import { CACHE_TTLS } from '../../../lib/utils'
 
 export async function GET(
-  request: NextRequest,
-  { params }: { params: Promise<{ fid: string }> }
+  request: NextRequest
 ) {
   try {
-    const { fid } = await params
+    const fid = request.nextUrl.searchParams.get('fid')
     const signer = request.nextUrl.searchParams.get('signer')
+    
+    if (!fid) {
+      return Response.json({ error: 'FID parameter required' }, { status: 400 })
+    }
     
     if (!signer) {
       return Response.json({ error: 'Signer parameter required' }, { status: 400 })
@@ -52,6 +55,7 @@ export async function GET(
       }
     })
   } catch (error) {
+    console.error('Error fetching signer stats:', error)
     return Response.json({ error: 'Failed to fetch signer stats' }, { status: 500 })
   }
 }
